Moved call contract to rpc
