+++ starknet-foundry/crates/cheatable-starknet/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod constants;
 pub mod state;
+pub mod rpc;
